requirements.txt
fastapi==0.115.8
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.2
pytest==8.0.1
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
loguru==0.7.2
fastapi-permissions==0.2.7
redis==4.6.0
aioredis==2.0.1
types-redis==4.6.0.20240106